x64: brgemm: other: eliminate implicit narrowing conversions
Group post-ops, transpose utils, and brdgmm buffer-size helpers.
Widen JIT-codegen-time host loop trip counts and dimension-derived
locals (nb, nb2, mb in post-ops; row_size_/tr_row_size_ chains in
jit_brgemm_copy_to_coarse_t) from int to dim_t, keeping narrowing
casts only at genuine hardware/register boundaries (masks, tails,
block sizes, Xbyak register indices). Route the remaining simd_w_
narrowing through the dim_t-safe xbyak_register_index() helper.

+++ src/cpu/x64/brgemm/brgemm.cpp
@@ -798,7 +798,7 @@ status_t brgemm_init_tiles(const brgemm_desc_t &brg, char palette[64]) {
         }
     }
 
-    buff->palette_id = amx::get_target_palette();
+    buff->palette_id = static_cast<uint8_t>(amx::get_target_palette());
 
     return status::success;
 }
